Refactor, memorialize the original functions while allowing for their adaptation
